+++ NEWS.md
@@ -1,6 +1,7 @@
 # idaifieldR 0.2.0 _2022-02-10_
 
 * Version number changed to 0.2, to reflect the dramatic changes (that have not yet taken place).
+* Faster `get_uid_list()` using `lapply()`
 * Added `idf_query()`, todo: will rely more on couchDB-queries in the future and therefore need to redo a lot of the package
 * Demo.Rmd updated to begin working on that.